+++ ephios/plugins/federation/templates/federation/external_event_list.html
@@ -2,11 +2,11 @@
 {% load i18n %}
 
 {% block title %}
-    {% translate "Shared events" %}
+    {% translate "External events" %}
 {% endblock %}
 
 {% block content %}
-    <h1>{% translate "Shared events" %}</h1>
+    <h1>{% translate "External events" %}</h1>
     {% for event in events %}
         <div class="list-group-item list-group-item-action p-0 d-flex flex-row">
             <div class="m-0 py-2 d-flex flex-column flex-lg-row-reverse justify-content-around flex-grow-0">
