+++ ephios/plugins/federation/templates/federation/invitecode_form.html
@@ -1,11 +1,12 @@
 {% extends "core/settings/settings_base.html" %}
+{% load i18n %}
 {% load crispy_forms_filters %}
 
 {% block settings_content %}
-    <h3>Add outgoing share</h3>
+    <h3>{% translate "Share events with someone else" %}</h3>
     <form method="post">
         {% csrf_token %}
         {{ form|crispy }}
-        <input type="submit" class="btn btn-primary" value="Add">
+        <input type="submit" class="btn btn-primary" value="{% translate "Invite" %}">
     </form>
 {% endblock %}
